Order the rows by hash
While there are no guarantees that the resulting JavaScript object stays
ordered by the key this appears to result in a deterministic and actually
ordered output.

Besides looking better to the human reader this also should improve import
performance by a bit, because new entries are inserted in the UNIQUE index
at the end, instead of being inserted all over the place.